Automatize generating the document gallery.
