+++ docs/gallery/logo.md
@@ -1,0 +1,10 @@
+# Logo
+
+This is an example how the pplt icon is created.
+
+![logo](images/logo_large.svg){: style="width: 100%"}
+
+```python
+--8<--  "docs/gallery/logo.py::89"
+```
+
