+++ docs/gallery/run.sh
@@ -2,3 +2,4 @@
 python cmaps.py
 python legend.py
 python colorbar.py
+python logo.py
