--- gallery/colorbar/README.md
@@ -1,14 +1,0 @@
-## Colorbar
-
-This is an example of using `prettypyplot.plot.colorbar`.
-
-.. warning::
-    This is not yet correctly implemented for `STYLE='minimal'`.
-
-```python
-.. include:: ../../gallery/colorbar/colorbar.py
-    :start-line: 7
-```
-
-![default](gallery/colorbar/default_plot_colorbar_right.png)
-![top](gallery/colorbar/default_plot_colorbar_top.png)
