--- gallery/legend/README.md
@@ -1,12 +1,0 @@
-## Legend
-
-This is an example of using `prettypyplot.plot.legend`.
-
-!!! warning
-    This is not yet correctly implemented for `STYLE='minimal'`.
-
-```python
---8<--  "gallery/legend/legend.py:7:"
-```
-
-![legend](/gallery/legend/default_default_plot_legend.png)
