--- gallery/subplots/README.md
@@ -1,10 +1,0 @@
-## Subplots
-
-This is an example of using `prettypyplot.subplots` module.
-
-```python
-.. include:: ../../gallery/subplots/subplots.py
-    :start-line: 7
-```
-
-![subplots](gallery/subplots/subplots_8figs.png)
